+++ src/rule.h
@@ -9,21 +9,43 @@
 #define rule_clear_flag(r, x) ((r)->flags &= ~(x))
 #define rule_set_flag(r, x)   ((r)->flags |=  (x))
 
+#define rule_effects_check_flag(e, x) ((e)->flags  &  (x))
+#define rule_effects_clear_flag(e, x) ((e)->flags &= ~(x))
+#define rule_effects_set_flag(e, x)   ((e)->flags |=  (x))
+
 enum rule_flags
 {
-    RULE_APP_VALID       = 1 <<  0,
-    RULE_TITLE_VALID     = 1 <<  1,
-    RULE_ROLE_VALID      = 1 <<  2,
-    RULE_SUBROLE_VALID   = 1 <<  3,
-    RULE_APP_EXCLUDE     = 1 <<  4,
-    RULE_TITLE_EXCLUDE   = 1 <<  5,
-    RULE_ROLE_EXCLUDE    = 1 <<  6,
-    RULE_SUBROLE_EXCLUDE = 1 <<  7,
-    RULE_FOLLOW_SPACE    = 1 <<  8,
-    RULE_OPACITY         = 1 <<  9,
-    RULE_LAYER           = 1 << 10,
-    RULE_ONE_SHOT        = 1 << 11,
-    RULE_ONE_SHOT_REMOVE = 1 << 12
+    RULE_APP_VALID       = 1 << 0,
+    RULE_TITLE_VALID     = 1 << 1,
+    RULE_ROLE_VALID      = 1 << 2,
+    RULE_SUBROLE_VALID   = 1 << 3,
+    RULE_APP_EXCLUDE     = 1 << 4,
+    RULE_TITLE_EXCLUDE   = 1 << 5,
+    RULE_ROLE_EXCLUDE    = 1 << 6,
+    RULE_SUBROLE_EXCLUDE = 1 << 7,
+    RULE_ONE_SHOT        = 1 << 8,
+    RULE_ONE_SHOT_REMOVE = 1 << 9
+};
+
+enum rule_effects_flag
+{
+    RULE_FOLLOW_SPACE = 1 << 0,
+    RULE_OPACITY      = 1 << 1,
+    RULE_LAYER        = 1 << 2,
+};
+
+struct rule_effects
+{
+    uint32_t did;
+    uint64_t sid;
+    float opacity;
+    int manage;
+    int sticky;
+    int mff;
+    int layer;
+    int fullscreen;
+    unsigned grid[6];
+    uint16_t flags;
 };
 
 struct rule
@@ -37,19 +59,12 @@ struct rule
     regex_t title_regex;
     regex_t role_regex;
     regex_t subrole_regex;
-    uint32_t did;
-    uint64_t sid;
-    float opacity;
-    int manage;
-    int sticky;
-    int mff;
-    int layer;
-    int fullscreen;
-    unsigned grid[6];
-    uint32_t flags;
+    struct rule_effects effects;
+    uint16_t flags;
 };
 
 void rule_serialize(FILE *rsp, struct rule *rule, int index);
+void rule_combine_effects(struct rule_effects *rule_effects, struct rule_effects *result);
 void rule_reapply_all(void);
 bool rule_reapply_by_index(int index);
 bool rule_reapply_by_label(char *label);