+++ src/window_manager.c
@@ -79,118 +79,125 @@ void window_manager_query_windows_for_displays(FILE *rsp)
     window_manager_query_windows_for_spaces(rsp, space_list, space_count);
 }
 
-void window_manager_apply_manage_rule_to_window(struct space_manager *sm, struct window_manager *wm, struct window *window, struct rule *rule, char *window_title, char *window_role, char *window_subrole)
+bool window_manager_rule_matches_window(struct rule *rule, struct window *window, char *window_title, char *window_role, char *window_subrole)
 {
     int regex_match_app = rule_check_flag(rule, RULE_APP_EXCLUDE) ? REGEX_MATCH_YES : REGEX_MATCH_NO;
-    if (regex_match(rule_check_flag(rule, RULE_APP_VALID), &rule->app_regex, window->application->name) == regex_match_app) return;
+    if (regex_match(rule_check_flag(rule, RULE_APP_VALID), &rule->app_regex, window->application->name) == regex_match_app) return false;
 
     int regex_match_title = rule_check_flag(rule, RULE_TITLE_EXCLUDE) ? REGEX_MATCH_YES : REGEX_MATCH_NO;
-    if (regex_match(rule_check_flag(rule, RULE_TITLE_VALID), &rule->title_regex, window_title) == regex_match_title) return;
+    if (regex_match(rule_check_flag(rule, RULE_TITLE_VALID), &rule->title_regex, window_title) == regex_match_title) return false;
 
     int regex_match_role = rule_check_flag(rule, RULE_ROLE_EXCLUDE) ? REGEX_MATCH_YES : REGEX_MATCH_NO;
-    if (regex_match(rule_check_flag(rule, RULE_ROLE_VALID), &rule->role_regex, window_role) == regex_match_role) return;
+    if (regex_match(rule_check_flag(rule, RULE_ROLE_VALID), &rule->role_regex, window_role) == regex_match_role) return false;
 
     int regex_match_subrole = rule_check_flag(rule, RULE_SUBROLE_EXCLUDE) ? REGEX_MATCH_YES : REGEX_MATCH_NO;
-    if (regex_match(rule_check_flag(rule, RULE_SUBROLE_VALID), &rule->subrole_regex, window_subrole) == regex_match_subrole) return;
+    if (regex_match(rule_check_flag(rule, RULE_SUBROLE_VALID), &rule->subrole_regex, window_subrole) == regex_match_subrole) return false;
 
-    if (rule->manage == RULE_PROP_ON) {
-        if (!rule_check_flag(rule, RULE_ROLE_VALID)    && !string_equals(window_role   , "AXWindow"))         return;
-        if (!rule_check_flag(rule, RULE_SUBROLE_VALID) && !string_equals(window_subrole, "AXStandardWindow")) return;
+    return true;
+}
 
+void window_manager_apply_manage_rule_effects_to_window(struct space_manager *sm, struct window_manager *wm, struct window *window, struct rule_effects *effects, char *window_title, char *window_role, char *window_subrole)
+{
+    if (effects->manage == RULE_PROP_ON) {
         window_rule_set_flag(window, WINDOW_RULE_MANAGED);
         window_manager_make_window_floating(sm, wm, window, false, true);
-    } else if (rule->manage == RULE_PROP_OFF) {
+    } else if (effects->manage == RULE_PROP_OFF) {
         window_rule_clear_flag(window, WINDOW_RULE_MANAGED);
         window_manager_make_window_floating(sm, wm, window, true, true);
     }
-
-    if (rule_check_flag(rule, RULE_ONE_SHOT)) {
-        rule_set_flag(rule, RULE_ONE_SHOT_REMOVE);
-    }
 }
 
-void window_manager_apply_rule_to_window(struct space_manager *sm, struct window_manager *wm, struct window *window, struct rule *rule, char *window_title, char *window_role, char *window_subrole)
+void window_manager_apply_rule_effects_to_window(struct space_manager *sm, struct window_manager *wm, struct window *window, struct rule_effects *effects, char *window_title, char *window_role, char *window_subrole)
 {
-    int regex_match_app = rule_check_flag(rule, RULE_APP_EXCLUDE) ? REGEX_MATCH_YES : REGEX_MATCH_NO;
-    if (regex_match(rule_check_flag(rule, RULE_APP_VALID), &rule->app_regex, window->application->name) == regex_match_app) return;
-
-    int regex_match_title = rule_check_flag(rule, RULE_TITLE_EXCLUDE) ? REGEX_MATCH_YES : REGEX_MATCH_NO;
-    if (regex_match(rule_check_flag(rule, RULE_TITLE_VALID), &rule->title_regex, window_title) == regex_match_title) return;
-
-    int regex_match_role = rule_check_flag(rule, RULE_ROLE_EXCLUDE) ? REGEX_MATCH_YES : REGEX_MATCH_NO;
-    if (regex_match(rule_check_flag(rule, RULE_ROLE_VALID), &rule->role_regex, window_role) == regex_match_role) return;
-
-    int regex_match_subrole = rule_check_flag(rule, RULE_SUBROLE_EXCLUDE) ? REGEX_MATCH_YES : REGEX_MATCH_NO;
-    if (regex_match(rule_check_flag(rule, RULE_SUBROLE_VALID), &rule->subrole_regex, window_subrole) == regex_match_subrole) return;
-
-    if (!window_rule_check_flag(window, WINDOW_RULE_MANAGED)) {
-        if (!rule_check_flag(rule, RULE_ROLE_VALID)    && !string_equals(window_role   , "AXWindow"))         return;
-        if (!rule_check_flag(rule, RULE_SUBROLE_VALID) && !string_equals(window_subrole, "AXStandardWindow")) return;
-    }
-
-    if (rule->sid || rule->did) {
+    if (effects->sid || effects->did) {
         if (!window_is_fullscreen(window) && !space_is_fullscreen(window_space(window->id))) {
-            uint64_t sid = rule->did ? display_space_id(rule->did) : rule->sid;
+            uint64_t sid = effects->did ? display_space_id(effects->did) : effects->sid;
             window_manager_send_window_to_space(sm, wm, window, sid, true);
-            if (rule_check_flag(rule, RULE_FOLLOW_SPACE) || rule->fullscreen == RULE_PROP_ON) {
+            if (rule_effects_check_flag(effects, RULE_FOLLOW_SPACE) || effects->fullscreen == RULE_PROP_ON) {
                 space_manager_focus_space(sid);
             }
         }
     }
 
-    if (rule->sticky == RULE_PROP_ON) {
+    if (effects->sticky == RULE_PROP_ON) {
         window_manager_make_window_sticky(sm, wm, window, true);
-    } else if (rule->sticky == RULE_PROP_OFF) {
+    } else if (effects->sticky == RULE_PROP_OFF) {
         window_manager_make_window_sticky(sm, wm, window, false);
     }
 
-    if (rule->mff == RULE_PROP_ON) {
+    if (effects->mff == RULE_PROP_ON) {
         window_rule_set_flag(window, WINDOW_RULE_MFF);
         window_rule_set_flag(window, WINDOW_RULE_MFF_VALUE);
-    } else if (rule->mff == RULE_PROP_OFF) {
+    } else if (effects->mff == RULE_PROP_OFF) {
         window_rule_set_flag(window, WINDOW_RULE_MFF);
         window_rule_clear_flag(window, WINDOW_RULE_MFF_VALUE);
     }
 
-    if (rule_check_flag(rule, RULE_LAYER)) {
-        window_manager_set_window_layer(window, rule->layer);
+    if (rule_effects_check_flag(effects, RULE_LAYER)) {
+        window_manager_set_window_layer(window, effects->layer);
     }
 
-    if (rule_check_flag(rule, RULE_OPACITY) && in_range_ii(rule->opacity, 0.0f, 1.0f)) {
-        window->opacity = rule->opacity;
-        window_manager_set_opacity(wm, window, rule->opacity);
+    if (rule_effects_check_flag(effects, RULE_OPACITY) && in_range_ii(effects->opacity, 0.0f, 1.0f)) {
+        window->opacity = effects->opacity;
+        window_manager_set_opacity(wm, window, effects->opacity);
     }
 
-    if (rule->fullscreen == RULE_PROP_ON) {
+    if (effects->fullscreen == RULE_PROP_ON) {
         AXUIElementSetAttributeValue(window->ref, kAXFullscreenAttribute, kCFBooleanTrue);
         window_rule_set_flag(window, WINDOW_RULE_FULLSCREEN);
     }
 
-    if (rule->grid[0] != 0 && rule->grid[1] != 0) {
-        window_manager_apply_grid(sm, wm, window, rule->grid[0], rule->grid[1], rule->grid[2], rule->grid[3], rule->grid[4], rule->grid[5]);
-    }
-
-    if (rule_check_flag(rule, RULE_ONE_SHOT)) {
-        rule_set_flag(rule, RULE_ONE_SHOT_REMOVE);
+    if (effects->grid[0] != 0 && effects->grid[1] != 0) {
+        window_manager_apply_grid(sm, wm, window, effects->grid[0], effects->grid[1], effects->grid[2], effects->grid[3], effects->grid[4], effects->grid[5]);
     }
 }
 
 void window_manager_apply_manage_rules_to_window(struct space_manager *sm, struct window_manager *wm, struct window *window, char *window_title, char *window_role, char *window_subrole, bool one_shot_rules)
 {
+    bool match = false;
+    struct rule_effects effects = {};
+
     for (int i = 0; i < buf_len(wm->rules); ++i) {
         if (one_shot_rules || !rule_check_flag(&wm->rules[i], RULE_ONE_SHOT)) {
-            window_manager_apply_manage_rule_to_window(sm, wm, window, &wm->rules[i], window_title, window_role, window_subrole);
+            if (window_manager_rule_matches_window(&wm->rules[i], window, window_title, window_role, window_subrole)) {
+                if (wm->rules[i].effects.manage == RULE_PROP_ON) {
+                    if (!rule_check_flag(&wm->rules[i], RULE_ROLE_VALID)    && !string_equals(window_role   , "AXWindow"))         continue;
+                    if (!rule_check_flag(&wm->rules[i], RULE_SUBROLE_VALID) && !string_equals(window_subrole, "AXStandardWindow")) continue;
+                }
+
+                match = true;
+                rule_combine_effects(&wm->rules[i].effects, &effects);
+
+                if (rule_check_flag(&wm->rules[i], RULE_ONE_SHOT)) rule_set_flag(&wm->rules[i], RULE_ONE_SHOT_REMOVE);
+            }
         }
     }
+
+    if (match) window_manager_apply_manage_rule_effects_to_window(sm, wm, window, &effects, window_title, window_role, window_subrole);
 }
 
 void window_manager_apply_rules_to_window(struct space_manager *sm, struct window_manager *wm, struct window *window, char *window_title, char *window_role, char *window_subrole, bool one_shot_rules)
 {
+    bool match = false;
+    struct rule_effects effects = {};
+
     for (int i = 0; i < buf_len(wm->rules); ++i) {
         if (one_shot_rules || !rule_check_flag(&wm->rules[i], RULE_ONE_SHOT)) {
-            window_manager_apply_rule_to_window(sm, wm, window, &wm->rules[i], window_title, window_role, window_subrole);
+            if (window_manager_rule_matches_window(&wm->rules[i], window, window_title, window_role, window_subrole)) {
+                if (!window_rule_check_flag(window, WINDOW_RULE_MANAGED)) {
+                    if (!rule_check_flag(&wm->rules[i], RULE_ROLE_VALID)    && !string_equals(window_role   , "AXWindow"))         continue;
+                    if (!rule_check_flag(&wm->rules[i], RULE_SUBROLE_VALID) && !string_equals(window_subrole, "AXStandardWindow")) continue;
+                }
+
+                match = true;
+                rule_combine_effects(&wm->rules[i].effects, &effects);
+
+                if (rule_check_flag(&wm->rules[i], RULE_ONE_SHOT)) rule_set_flag(&wm->rules[i], RULE_ONE_SHOT_REMOVE);
+            }
         }
     }
+
+    if (match) window_manager_apply_rule_effects_to_window(sm, wm, window, &effects, window_title, window_role, window_subrole);
 }
 
 void window_manager_set_focus_follows_mouse(struct window_manager *wm, enum ffm_mode mode)